fix benchmarks and cookbooks problems

+++ benchmarks/solitary_wave/solitary_wave.cc
@@ -284,7 +284,7 @@ namespace aspect
           virtual void vector_value (const Point< dim > &p,
                                      Vector< double >   &values) const
           {
-            double index = static_cast<int>((p[dim-1]-delta_)/max_z_ * (initial_pressure_.size()-1));
+            unsigned int index = static_cast<int>((p[dim-1]-delta_)/max_z_ * (initial_pressure_.size()-1));
             if (p[dim-1]-delta_ < 0)
               index = 0;
             else if (p[dim-1]-delta_ > max_z_)
